pyproject.toml
[project]
name = "stravation"
version = "0.1.0"
description = "Strava → Notion + Prévision + Google Calendar (rappels matin) — clean & minimal"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pendulum>=3.0",
  "pydantic>=2.7",
  "httpx>=0.27",
  "google-api-python-client>=2.149.0",
  "google-auth>=2.34.0",
  "google-auth-oauthlib>=1.2.1",
  "notion-client>=2.2.1",
  "sqlite-utils>=3.36",
  "python-dotenv>=1.0.1",
]

[project.scripts]
stravation = "stravation.cli:app"   # 👈 nouveau nom de commande
